feat(ui): add confirmation to delete actions [EE-4612]
closes [EE-4612]

adds confirmation to:
- image delete tag
- image view delete image
- edge groups delete

rewords the confirmations for:
- images delete and force delete
- edge jobs

+++ app/edge/views/edge-groups/edgeGroupsView/edgeGroupsViewController.js
@@ -1,4 +1,5 @@
 import _ from 'lodash-es';
+import { confirmDelete } from '@@/modals/confirm';
 
 export class EdgeGroupsController {
   /* @ngInject */
@@ -26,6 +27,10 @@ export class EdgeGroupsController {
   }
 
   async removeActionAsync(selectedItems) {
+    if (!(await confirmDelete('Do you want to remove the selected Edge Group(s)?'))) {
+      return;
+    }
+
     for (let item of selectedItems) {
       try {
         await this.EdgeGroupService.remove(item.Id);
